feat: add study metadata save, status, and submit controllers
- Implemented createStudyMetadataSaveController for saving study metadata with success and error handling.
- Added createMetadataStatusController to manage and display metadata sync and citation health statuses.
- Introduced createStudyMetadataSubmitController to handle form submission for study metadata.
- Developed project hints functionality to provide inline help for form fields.
- Created settings management for global library and backend monitoring settings.
- Added utility functions for managing beginner help mode and rendering field hints.
- Included new files for handling project settings and library information updates.
