+++ test_exit.txt
@@ -1,0 +1,1 @@
+0
